Fix hitting steps rather than climbing
Issue was that when we hit a wall and we check for steps, we were looking for them starting from the current position rather than the point of collision. Meaning that if your current velocity was higher than the minimum step depth, it wouldn't find the step and think it's a wall instead.