Add Pseudo Steps
Pseudo steps are small patches of terrain that is too steep for you to walk on, but they are smaller than a step you could normally walk over, so when enabled you'll walk over them regardless if they're too steep.
Makes the movement smoother at low velocities and more reliable on rough terrain.